Moving added functionality from `Logic/CoAuthorPlus.php` to `Logic/CoAuthorPlusDataFixer.php`
